Split sample components into separate files
